backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
sqlalchemy==2.0.27
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1 